+++ src/main/java/com/gogoyang/rpgapi/meta/honor/dao/HonorDao.java
@@ -31,5 +31,5 @@ public interface HonorDao{
      * type
      * @return
      */
-    Integer sumHonor(Map qIn);
+    Double sumHonor(Map qIn);
 }
