+++ src/main/java/com/gogoyang/rpgapi/meta/honor/service/IHonorService.java
@@ -31,5 +31,5 @@ public interface IHonorService {
      * type
      * @return
      */
-    Integer sumHonor(Map qIn);
+    Double sumHonor(Map qIn);
 }
